Reworked parser: FOR now mostly works
Parser is now in a class, to decompose subsections.  This turned out not to be as useful as I'd hoped.
Some new default modifiers were added for redcode-94 access modes which may be correct; I think they're correct, but mostly it was to stop parse errors.
EQU has a known bug where it is still a variable instead of a rewrite rule, which fails several warriors on Koenigstuhl.  People were using it to clone entire opcodes, so it needs reworking.
Added some exception handling to the rule eval stuff.