general: Handle non-UTF8 encoded warriors
Request more lenient error handling while reading warriors to
prevent failing loading old files that might have comments that are
not UTF-8 encoded.

'surrogateescape' was chosen, instead of 'replace' or 'ignore'
so that the transformation is lossless :incorrect characters are mapped
to the unicode private use area. It means that theoretically the
encoding could be guessed and corrected for relevant fields like `name`
and `author` after reading (not done here).